requirements.txt
Bootstrap_Flask>=2.2.0
Flask>=2.2.5
Pillow>=8.4.0
numpy>=1.21.2
colorthief>=0.3.0
